Don't need scheduler.
The purpose of a scheduler is to schedule tasks, in some order. The
reason I went down the route of writing my own scheduler was because I
thought that we need to guarantee I/O operations happen in a specific
order, to guarantee determinism. I suppose at some low level this may be
required but generally it shouldn't be the responsibility of the
simulator. The client knows which I/O operations need to be
synchronised, not the simulator, and so it is the client's
responsibility to synchronise them.